requirements.txt
flask
python-dotenv
gtts
pygame
psycopg2-binary
